requirements.txt
django_otp==0.8.1
django_heroku==0.3.1
Django==3.0.4
razorpay==1.2.0
requests==2.20.1
PyJWT==1.7.1
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.11.3
google-auth-httplib2==0.0.3
googleapis-common-protos==1.51.0
psycopg2==2.8.3
setuptools==46.1.3
gunicorn==20.0.4